Module google_pubsublite1::api

Structs

AdminMethods
A builder providing access to all methods supported on admin resources. It is not used directly, but through the PubsubLite hub.

AdminProjectLocationSubscriptionCreateCall
Creates a new subscription.

AdminProjectLocationSubscriptionDeleteCall
Deletes the specified subscription.

AdminProjectLocationSubscriptionGetCall
Returns the subscription configuration.

AdminProjectLocationSubscriptionListCall
Returns the list of subscriptions for the given project.

AdminProjectLocationSubscriptionPatchCall
Updates properties of the specified subscription.

AdminProjectLocationTopicCreateCall
Creates a new topic.

AdminProjectLocationTopicDeleteCall
Deletes the specified topic.

AdminProjectLocationTopicGetCall
Returns the topic configuration.

AdminProjectLocationTopicGetPartitionCall
Returns the partition information for the requested topic.

AdminProjectLocationTopicListCall
Returns the list of topics for the given project.

AdminProjectLocationTopicPatchCall
Updates properties of the specified topic.

AdminProjectLocationTopicSubscriptionListCall
Lists the subscriptions attached to the specified topic.

Capacity
The throughput capacity configuration for each partition.

ComputeHeadCursorRequest
Compute the current head cursor for a partition.

ComputeHeadCursorResponse
Response containing the head cursor for the requested topic and partition.

ComputeMessageStatsRequest
Compute statistics about a range of messages in a given topic and partition.

ComputeMessageStatsResponse
Response containing stats for messages in the requested topic and partition.

Cursor
A cursor that describes the position of a message within a topic partition.

CursorMethods
A builder providing access to all methods supported on cursor resources. It is not used directly, but through the PubsubLite hub.

CursorProjectLocationSubscriptionCursorListCall
Returns all committed cursor information for a subscription.

DeliveryConfig
The settings for a subscription's message delivery.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

ListPartitionCursorsResponse
Response for ListPartitionCursors

ListSubscriptionsResponse
Response for ListSubscriptions.

ListTopicSubscriptionsResponse
Response for ListTopicSubscriptions.

ListTopicsResponse
Response for ListTopics.

PartitionConfig
The settings for a topic's partitions.

PartitionCursor
A pair of a Cursor and the partition it is for.

PubsubLite
Central instance to access all PubsubLite related resource activities

RetentionConfig
The settings for a topic's message retention.

Subscription
Metadata about a subscription resource.

Topic
Metadata about a topic resource.

TopicPartitions
Response for GetTopicPartitions.

TopicStatMethods
A builder providing access to all methods supported on topicStat resources. It is not used directly, but through the PubsubLite hub.

TopicStatProjectLocationTopicComputeHeadCursorCall
Compute the head cursor for the partition. The head cursor's offset is guaranteed to be less than or equal to all messages which have not yet been acknowledged as published, and greater than the offset of any message whose publish has already been acknowledged. It is zero if there have never been messages in the partition.

TopicStatProjectLocationTopicComputeMessageStatCall
Compute statistics about a range of messages in a given topic and partition.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.